Allow configuration of Calico bgp and encapsulation values
To enable Calico BGP mode we need to set bgp and encapsulation values.
The chart allows overriding of the bgp value already, but encapsulation
is within a list, which is concatenated if changes are attempted in the
overrides field.

So, here we add configuration for both values and document how to enable
the full-mesh mode we will use.