Sort node-labels whilst iterating over keys (#321)
Sorts the node-labels whilst iterating through the keys of the
un-ordered dict. This prevents a machine deployment flipping between n
machine-sets.

Currently, each evaluation of the template could result in a different
label order. Since the entire resource is hashed, this results in
"upgrades" back and forth, depending on where we land in the unordered
dict.

By sorting the keys, we at least ensure that the results are idempotent
between helm template (or equiv.) runs.

I'm sure there's other places we rely on range implicitly being ordered,
but doing a quick grep through, the other uses of range in
openstack-cluster/helpers.tpl appear to not require this from a glance.