Prepend default to GlobalNetPolicy manifest name
The manifest source targets {{ cluster-name }}, however the operator
prepends default.{{ cluster-name}}. This works fine until there's a diff
(e.g. the version changes in the annotation). Then the manifest will
fail to apply/patch because K8s requires a delete and recreate for name
changes.

We've seen some clusters which have sat retrying for 20K+ attempts, but
because the failure is an annotation the cluster continues fine so it's
probably been missed for a while